+++ tobago-example/tobago-example-demo/src/main/webapp/content/080-sheet/32-action/Sheet_Action.xhtml
@@ -1,0 +1,39 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<!--
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+-->
+
+<ui:composition template="/main.xhtml"
+                xmlns="http://www.w3.org/1999/xhtml"
+                xmlns:tc="http://myfaces.apache.org/tobago/component"
+                xmlns:ui="http://xmlns.jcp.org/jsf/facelets"
+                xmlns:f="http://xmlns.jcp.org/jsf/core">
+
+  <tc:section label="Action in a sheet">
+
+    <tc:sheet id="sheet" value="#{sheetController.solarList}" var="object" markup="small">
+      <tc:column label="Name">
+        <tc:out value="#{object.name}" id="name" labelLayout="skip"/>
+      </tc:column>
+      <tc:column label="Action">
+        <tc:link action="#{sheetActionController.deleteAction(object)}"
+                 image="bi-trash" id="action" labelLayout="skip"/>
+      </tc:column>
+    </tc:sheet>
+
+  </tc:section>
+</ui:composition>
